Criando modulo para eventos.
